Add Task db model to track the progress of generating a deck.
Creating a deck and auto generating the flashcards through openai is slow. Instead of
waiting for the process to complete then respond to the user, the server returns immediately
and returns a task id. The server asynchronously runs the task of building the deck without awaiting
the results, and the user can use the task id to query the status of creating the deck.